Change DAOManager layer back to returning instances of DAO interfaces with separate classes for Tapestry entities layer
I think we are going to need this after all, and it's a cleaner separation of layers in the end, even though it needs
more conversion code.

This way the DAO objects such as PolicyDAOImpl can contain datastore specific stuff such as timestamps, tracking of
which fields have changed, etc.