Remove attempt to set column metadata on subcolumns
Can't do this it seems!

Only supported for columns in normal column families, not for subcolumns in supercolumn families.
It seems that if you use it on supercolumn families, it tries applies it to the supercolumns,
and validation_class in the metadata conflicts with the content actually being a list of subcolumns,
which causes failure to start up. Nice.

Doesn't really matter, we have strong type enforcement and validation via our Schema abstraction and query mechanism.