Implement user vote salary creation and retrieval
Changed design slightly: we now have a single row containing vote salary history for the whole system,
and users' individual salary histories/balances are just a tail of this list based on their registration date.

Implemented auto-creation of these salary records as needed (when we try to retrieve info),
plus config of vote salary amount, salary frequency, and (if it's weekly) the day of the week to use.
(First salary payment will occur on the date the system starts; subsequent ones will be on the specified day.)

TODO: unit tests