Fix list construction to work in javac compiler
For some reason I can't get the CollectionUtils.list idiom to work with these generic values with the single-arg method
with javac, even though it works with the multiple-arg method and the single-arg one works with the eclipse compiler.

Give up, just make a local ArrayList when there's only one element.