Implement setting list of cassandra hosts and username/password from servlet context parameters
See params in web.xml, defaulting to the same localhost:9160 as before.

The connection to cassandra is now authenticated (in webapp mode only - test uses AllowAll as before).
Authentication is enabled by setting SimpleAuthenticator in cassandra.yaml, and providing it with passwd.properties.

A default password is provided for development and hardcoded in web.xml and the test scripts;
in production we'll just override this from tomcat servlet context config and specify manually when running initial cluster setup.

Also tweaked various other Hector connection parameters.