Final changes before testing & push, new user page now works, creates an activation code (statically set in NewUser.java to testCode) and stores that in the DB.
Flow should then transfer to a URL /users/activateuser which expects two parameters on the request, code (activation code) and email that user used to register.

This then prompts for other info, all fields are required at the moment as ValueImpl blocks null values (is this expected to change, am I doing it wrong?). Ta da, you now have a user!

Roles in activateUser are optional, defaults will be set if none are provided.